requirements.txt
boto3==1.28.84
botocore==1.31.84
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
et-xmlfile==1.1.0
google-api-core==2.14.0
google-auth==2.23.4
google-cloud-core==2.3.3
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
idna==3.4
jmespath==1.0.1
numpy==1.26.1
openpyxl==3.1.2
pandas==2.1.3
protobuf==4.25.0
psycopg2==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyodbc==5.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
s3transfer==0.7.0
six==1.16.0
tzdata==2023.3
urllib3==1.26.18
mysql.connector==2.2.9